refactored exceptions
