+++ src/gt4py/frontend/gtscript_frontend.py
@@ -37,78 +37,7 @@ from gtc import utils as gtc_utils
 
 from .base import Frontend, register
 
-
-class GTScriptSyntaxError(gt_definitions.GTSyntaxError):
-    def __init__(self, message, *, loc=None):
-        super().__init__(message, frontend=GTScriptFrontend.name)
-        self.loc = loc
-
-
-class GTScriptSymbolError(GTScriptSyntaxError):
-    def __init__(self, name, message=None, *, loc=None):
-        if message is None:
-            if loc is None:
-                message = "Unknown symbol '{name}' symbol".format(name=name)
-            else:
-                message = (
-                    "Unknown symbol '{name}' symbol in '{scope}' (line: {line}, col: {col})".format(
-                        name=name, scope=loc.scope, line=loc.line, col=loc.column
-                    )
-                )
-        super().__init__(message, loc=loc)
-        self.name = name
-
-
-class GTScriptDefinitionError(GTScriptSyntaxError):
-    def __init__(self, name, value, message=None, *, loc=None):
-        if message is None:
-            if loc is None:
-                message = "Invalid definition for '{name}' symbol".format(name=name)
-            else:
-                message = "Invalid definition for '{name}' symbol in '{scope}' (line: {line}, col: {col})".format(
-                    name=name, scope=loc.scope, line=loc.line, col=loc.column
-                )
-        super().__init__(message, loc=loc)
-        self.name = name
-        self.value = value
-
-
-class GTScriptValueError(GTScriptDefinitionError):
-    def __init__(self, name, value, message=None, *, loc=None):
-        if message is None:
-            if loc is None:
-                message = "Invalid value for '{name}' symbol ".format(name=name)
-            else:
-                message = (
-                    "Invalid value for '{name}' in '{scope}' (line: {line}, col: {col})".format(
-                        name=name, scope=loc.scope, line=loc.line, col=loc.column
-                    )
-                )
-        super().__init__(name, value, message, loc=loc)
-
-
-class GTScriptDataTypeError(GTScriptSyntaxError):
-    def __init__(self, name, data_type, message=None, *, loc=None):
-        if message is None:
-            if loc is None:
-                message = "Invalid data type for '{name}' numeric symbol ".format(name=name)
-            else:
-                message = "Invalid data type for '{name}' numeric symbol in '{scope}' (line: {line}, col: {col})".format(
-                    name=name, scope=loc.scope, line=loc.line, col=loc.column
-                )
-        super().__init__(message, loc=loc)
-        self.name = name
-        self.data_type = data_type
-
-
-class GTScriptAssertionError(gt_definitions.GTSpecificationError):
-    def __init__(self, source, *, loc=None):
-        if loc:
-            message = f"Assertion failed at line {loc.line}, col {loc.column}:\n{source}"
-        else:
-            message = f"Assertion failed.\n{source}"
-        super().__init__(message)
-        self.loc = loc
+from .exceptions import GTScriptSyntaxError, GTScriptAssertionError, GTScriptDefinitionError, GTScriptValueError, GTScriptSymbolError, GTScriptDataTypeError
 
 
 class AssertionChecker(ast.NodeTransformer):
